added vnr task file
